pyproject.toml
[project]
name = "mlops-assignment"
version = "0.1.0"
description = ""
authors = [{ name = "Alex, Thadenn" }]
readme = "README.md"
requires-python = "3.10.18"
dependencies = [
    "pandas==2.1.4",
    "pycaret[analysis,models,tuners,mlops]==3.3.2",
    "mlflow==3.3.2",
    "hydra-core==1.3.2",
    "scikit-learn==1.4.2",
    "openpyxl==3.1.5",
    "streamlit==1.49.0",
    "dvc==3.62.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
